simulation/requirements.txt
certifi==2020.12.5
chardet==4.0.0
docplex==2.20.204
mysql-connector-python==8.0.23
numpy==1.20.2
pandas==1.2.3
plotly==4.14.3
protobuf==3.15.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
simpy==4.0.1
six==1.15.0
urllib3==1.26.4
